Move draw parameters synchronization to ops/draw.rs
