+++ src/draw_parameters.rs
@@ -1,5 +1,4 @@
 use gl;
-use context;
 use context::GlVersion;
 use version::Api;
 
@@ -476,292 +475,4 @@ pub fn validate(display: &Display, params: &DrawParameters) -> Result<(), DrawEr
     }
 
     Ok(())
-}
-
-/// Synchronizes the parameters with the current ctxt.state.
-pub fn sync(params: &DrawParameters, ctxt: &mut context::CommandContext,
-            surface_dimensions: (u32, u32))
-{
-    // depth function
-    match params.depth_test {
-        DepthTest::Overwrite => unsafe {
-            if ctxt.state.enabled_depth_test {
-                ctxt.gl.Disable(gl::DEPTH_TEST);
-                ctxt.state.enabled_depth_test = false;
-            }
-        },
-        depth_function => unsafe {
-            let depth_function = depth_function.to_glenum();
-            if ctxt.state.depth_func != depth_function {
-                ctxt.gl.DepthFunc(depth_function);
-                ctxt.state.depth_func = depth_function;
-            }
-            if !ctxt.state.enabled_depth_test {
-                ctxt.gl.Enable(gl::DEPTH_TEST);
-                ctxt.state.enabled_depth_test = true;
-            }
-        }
-    }
-
-    // depth mask
-    if params.depth_write != ctxt.state.depth_mask {
-        unsafe {
-            ctxt.gl.DepthMask(if params.depth_write { gl::TRUE } else { gl::FALSE });
-        }
-        ctxt.state.depth_mask = params.depth_write;
-    }
-
-    // depth range
-    if params.depth_range != ctxt.state.depth_range {
-        unsafe {
-            ctxt.gl.DepthRange(params.depth_range.0 as f64, params.depth_range.1 as f64);
-        }
-        ctxt.state.depth_range = params.depth_range;
-    }
-
-    // blending function
-    let blend_factors = match params.blending_function {
-        Some(BlendingFunction::AlwaysReplace) => unsafe {
-            if ctxt.state.enabled_blend {
-                ctxt.gl.Disable(gl::BLEND);
-                ctxt.state.enabled_blend = false;
-            }
-            None
-        },
-        Some(BlendingFunction::Min) => unsafe {
-            if ctxt.state.blend_equation != gl::MIN {
-                ctxt.gl.BlendEquation(gl::MIN);
-                ctxt.state.blend_equation = gl::MIN;
-            }
-            if !ctxt.state.enabled_blend {
-                ctxt.gl.Enable(gl::BLEND);
-                ctxt.state.enabled_blend = true;
-            }
-            None
-        },
-        Some(BlendingFunction::Max) => unsafe {
-            if ctxt.state.blend_equation != gl::MAX {
-                ctxt.gl.BlendEquation(gl::MAX);
-                ctxt.state.blend_equation = gl::MAX;
-            }
-            if !ctxt.state.enabled_blend {
-                ctxt.gl.Enable(gl::BLEND);
-                ctxt.state.enabled_blend = true;
-            }
-            None
-        },
-        Some(BlendingFunction::Addition { source, destination }) => unsafe {
-            if ctxt.state.blend_equation != gl::FUNC_ADD {
-                ctxt.gl.BlendEquation(gl::FUNC_ADD);
-                ctxt.state.blend_equation = gl::FUNC_ADD;
-            }
-            if !ctxt.state.enabled_blend {
-                ctxt.gl.Enable(gl::BLEND);
-                ctxt.state.enabled_blend = true;
-            }
-            Some((source, destination))
-        },
-        Some(BlendingFunction::Subtraction { source, destination }) => unsafe {
-            if ctxt.state.blend_equation != gl::FUNC_SUBTRACT {
-                ctxt.gl.BlendEquation(gl::FUNC_SUBTRACT);
-                ctxt.state.blend_equation = gl::FUNC_SUBTRACT;
-            }
-            if !ctxt.state.enabled_blend {
-                ctxt.gl.Enable(gl::BLEND);
-                ctxt.state.enabled_blend = true;
-            }
-            Some((source, destination))
-        },
-        Some(BlendingFunction::ReverseSubtraction { source, destination }) => unsafe {
-            if ctxt.state.blend_equation != gl::FUNC_REVERSE_SUBTRACT {
-                ctxt.gl.BlendEquation(gl::FUNC_REVERSE_SUBTRACT);
-                ctxt.state.blend_equation = gl::FUNC_REVERSE_SUBTRACT;
-            }
-            if !ctxt.state.enabled_blend {
-                ctxt.gl.Enable(gl::BLEND);
-                ctxt.state.enabled_blend = true;
-            }
-            Some((source, destination))
-        },
-        _ => None
-    };
-    if let Some((source, destination)) = blend_factors {
-        let source = source.to_glenum();
-        let destination = destination.to_glenum();
-
-        if ctxt.state.blend_func != (source, destination) {
-            unsafe { ctxt.gl.BlendFunc(source, destination) };
-            ctxt.state.blend_func = (source, destination);
-        }
-    };
-
-    // line width
-    if let Some(line_width) = params.line_width {
-        if ctxt.state.line_width != line_width {
-            unsafe {
-                ctxt.gl.LineWidth(line_width);
-                ctxt.state.line_width = line_width;
-            }
-        }
-    }
-
-    // point size
-    if let Some(point_size) = params.point_size {
-        if ctxt.state.point_size != point_size {
-            unsafe {
-                ctxt.gl.PointSize(point_size);
-                ctxt.state.point_size = point_size;
-            }
-        }
-    }
-
-    // back-face culling
-    // note: we never change the value of `glFrontFace`, whose default is GL_CCW
-    //  that's why `CullClockWise` uses `GL_BACK` for example
-    match params.backface_culling {
-        BackfaceCullingMode::CullingDisabled => unsafe {
-            if ctxt.state.enabled_cull_face {
-                ctxt.gl.Disable(gl::CULL_FACE);
-                ctxt.state.enabled_cull_face = false;
-            }
-        },
-        BackfaceCullingMode::CullCounterClockWise => unsafe {
-            if !ctxt.state.enabled_cull_face {
-                ctxt.gl.Enable(gl::CULL_FACE);
-                ctxt.state.enabled_cull_face = true;
-            }
-            if ctxt.state.cull_face != gl::FRONT {
-                ctxt.gl.CullFace(gl::FRONT);
-                ctxt.state.cull_face = gl::FRONT;
-            }
-        },
-        BackfaceCullingMode::CullClockWise => unsafe {
-            if !ctxt.state.enabled_cull_face {
-                ctxt.gl.Enable(gl::CULL_FACE);
-                ctxt.state.enabled_cull_face = true;
-            }
-            if ctxt.state.cull_face != gl::BACK {
-                ctxt.gl.CullFace(gl::BACK);
-                ctxt.state.cull_face = gl::BACK;
-            }
-        },
-    }
-
-    // polygon mode
-    unsafe {
-        let polygon_mode = params.polygon_mode.to_glenum();
-        if ctxt.state.polygon_mode != polygon_mode {
-            ctxt.gl.PolygonMode(gl::FRONT_AND_BACK, polygon_mode);
-            ctxt.state.polygon_mode = polygon_mode;
-        }
-    }
-
-    // multisampling
-    if ctxt.state.enabled_multisample != params.multisampling {
-        unsafe {
-            if params.multisampling {
-                ctxt.gl.Enable(gl::MULTISAMPLE);
-                ctxt.state.enabled_multisample = true;
-            } else {
-                ctxt.gl.Disable(gl::MULTISAMPLE);
-                ctxt.state.enabled_multisample = false;
-            }
-        }
-    }
-
-    // dithering
-    if ctxt.state.enabled_dither != params.dithering {
-        unsafe {
-            if params.dithering {
-                ctxt.gl.Enable(gl::DITHER);
-                ctxt.state.enabled_dither = true;
-            } else {
-                ctxt.gl.Disable(gl::DITHER);
-                ctxt.state.enabled_dither = false;
-            }
-        }
-    }
-
-    // viewport
-    if let Some(viewport) = params.viewport {
-        assert!(viewport.width <= ctxt.capabilities.max_viewport_dims.0 as u32,
-                "Viewport dimensions are too large");
-        assert!(viewport.height <= ctxt.capabilities.max_viewport_dims.1 as u32,
-                "Viewport dimensions are too large");
-
-        let viewport = (viewport.left as gl::types::GLint, viewport.bottom as gl::types::GLint,
-                        viewport.width as gl::types::GLsizei,
-                        viewport.height as gl::types::GLsizei);
-
-        if ctxt.state.viewport != Some(viewport) {
-            unsafe { ctxt.gl.Viewport(viewport.0, viewport.1, viewport.2, viewport.3); }
-            ctxt.state.viewport = Some(viewport);
-        }
-
-    } else {
-        assert!(surface_dimensions.0 <= ctxt.capabilities.max_viewport_dims.0 as u32,
-                "Viewport dimensions are too large");
-        assert!(surface_dimensions.1 <= ctxt.capabilities.max_viewport_dims.1 as u32,
-                "Viewport dimensions are too large");
-
-        let viewport = (0, 0, surface_dimensions.0 as gl::types::GLsizei,
-                        surface_dimensions.1 as gl::types::GLsizei);
-
-        if ctxt.state.viewport != Some(viewport) {
-            unsafe { ctxt.gl.Viewport(viewport.0, viewport.1, viewport.2, viewport.3); }
-            ctxt.state.viewport = Some(viewport);
-        }
-    }
-
-    // scissor
-    if let Some(scissor) = params.scissor {
-        let scissor = (scissor.left as gl::types::GLint, scissor.bottom as gl::types::GLint,
-                       scissor.width as gl::types::GLsizei,
-                       scissor.height as gl::types::GLsizei);
-
-        unsafe {
-            if ctxt.state.scissor != Some(scissor) {
-                ctxt.gl.Scissor(scissor.0, scissor.1, scissor.2, scissor.3);
-                ctxt.state.scissor = Some(scissor);
-            }
-
-            if !ctxt.state.enabled_scissor_test {
-                ctxt.gl.Enable(gl::SCISSOR_TEST);
-                ctxt.state.enabled_scissor_test = true;
-            }
-        }
-    } else {
-        unsafe {
-            if ctxt.state.enabled_scissor_test {
-                ctxt.gl.Disable(gl::SCISSOR_TEST);
-                ctxt.state.enabled_scissor_test = false;
-            }
-        }
-    }
-
-    // draw_primitives
-    if ctxt.state.enabled_rasterizer_discard == params.draw_primitives {
-        if ctxt.version >= &GlVersion(Api::Gl, 3, 0) {
-            if params.draw_primitives {
-                unsafe { ctxt.gl.Disable(gl::RASTERIZER_DISCARD); }
-                ctxt.state.enabled_rasterizer_discard = false;
-            } else {
-                unsafe { ctxt.gl.Enable(gl::RASTERIZER_DISCARD); }
-                ctxt.state.enabled_rasterizer_discard = true;
-            }
-
-        } else if ctxt.extensions.gl_ext_transform_feedback {
-            if params.draw_primitives {
-                unsafe { ctxt.gl.Disable(gl::RASTERIZER_DISCARD_EXT); }
-                ctxt.state.enabled_rasterizer_discard = false;
-            } else {
-                unsafe { ctxt.gl.Enable(gl::RASTERIZER_DISCARD_EXT); }
-                ctxt.state.enabled_rasterizer_discard = true;
-            }
-
-        } else {
-            unreachable!();
-        }
-
-    }
 }
